requirements.txt
pandas~=2.2.3
kagglehub~=0.3.12
numpy~=2.0.1
tqdm~=4.67.1
torch~=2.5.1+cu121
transformers~=4.29.0
scikit-learn~=1.6.1
ipython~=8.36.0
requests~=2.32.3
hdbscan~=0.8.40
sentence-transformers~=2.2.2
umap-learn~=0.5.7
matplotlib~=3.10.3
seaborn~=0.13.2
wordcloud~=1.9.4
datasets~=3.6.0
imbalanced-learn~=0.13.0
torchvision~=0.20.1+cu121
joblib~=1.4.2
xgboost~=3.0.2
skl2onnx~=1.19.1
textblob~=0.19.0
onnxruntime-gpu==1.22.0